+++ packages/gatsby-source-contentful/src/gatsby-node.js
@@ -35,7 +35,16 @@ exports.onPreBootstrap = validateOptions
  */
 
 exports.sourceNodes = async (
-  { actions, getNode, getNodes, createNodeId, store, cache, reporter },
+  {
+    actions,
+    getNode,
+    getNodes,
+    createNodeId,
+    store,
+    cache,
+    getCache,
+    reporter,
+  },
   pluginOptions
 ) => {
   const { createNode, deleteNode, touchNode, setPluginStatus } = actions
@@ -243,6 +252,7 @@ exports.sourceNodes = async (
       createNodeId,
       store,
       cache,
+      getCache,
       getNodes,
       reporter,
     })