+++ packages/gatsby/src/redux/run-sift.js
@@ -3,7 +3,7 @@ const { default: sift } = require(`sift`)
 const _ = require(`lodash`)
 const { prepareRegex } = require(`../utils/prepare-regex`)
 const { makeRe } = require(`micromatch`)
-const { getValueAt } = require(`../utils/get-value-at`)
+import { getValueAt } from "../utils/get-value-at"
 const {
   toDottedFields,
   objectToDottedField,