+++ packages/gatsby/src/utils/webpack-utils.js
@@ -5,12 +5,13 @@ const flexbugs = require(`postcss-flexbugs-fixes`)
 const TerserPlugin = require(`terser-webpack-plugin`)
 const MiniCssExtractPlugin = require(`mini-css-extract-plugin`)
 const OptimizeCssAssetsPlugin = require(`optimize-css-assets-webpack-plugin`)
+const ReactRefreshWebpackPlugin = require(`@pmmmwh/react-refresh-webpack-plugin`)
 const isWsl = require(`is-wsl`)
 
 const GatsbyWebpackStatsExtractor = require(`./gatsby-webpack-stats-extractor`)
 const GatsbyWebpackEslintGraphqlSchemaReload = require(`./gatsby-webpack-eslint-graphql-schema-reload-plugin`)
 
-const builtinPlugins = require(`./webpack-plugins`)
+import { builtinPlugins } from "./webpack-plugins"
 const eslintConfig = require(`./eslint-config`)
 
 type LoaderSpec = string | { loader: string, options?: Object }
@@ -131,7 +132,7 @@ module.exports = async ({
   const makeExternalOnly = (original: RuleFactory<*>) => (
     options = {}
   ): Rule => {
-    let rule = original(options)
+    const rule = original(options)
     rule.include = vendorRegex
     return rule
   }
@@ -139,7 +140,7 @@ module.exports = async ({
   const makeInternalOnly = (original: RuleFactory<*>) => (
     options = {}
   ): Rule => {
-    let rule = original(options)
+    const rule = original(options)
     rule.exclude = vendorRegex
     return rule
   }
@@ -306,7 +307,7 @@ module.exports = async ({
    * and packages that depend on `gatsby`
    */
   {
-    let js = ({ modulesThatUseGatsby = [], ...options } = {}) => {
+    const js = ({ modulesThatUseGatsby = [], ...options } = {}) => {
       return {
         test: /\.(js|mjs|jsx)$/,
         include: modulePath => {
@@ -341,7 +342,7 @@ module.exports = async ({
    * Excludes modules that use Gatsby since the `rules.js` already transpiles those
    */
   {
-    let dependencies = ({ modulesThatUseGatsby = [] } = {}) => {
+    const dependencies = ({ modulesThatUseGatsby = [] } = {}) => {
       const jsOptions = {
         babelrc: false,
         configFile: false,
@@ -390,7 +391,7 @@ module.exports = async ({
   }
 
   {
-    let eslint = schema => {
+    const eslint = schema => {
       return {
         enforce: `pre`,
         test: /\.jsx?$/,
@@ -608,6 +609,11 @@ module.exports = async ({
       },
     }
   ) => new OptimizeCssAssetsPlugin(options)
+
+  plugins.fastRefresh = () =>
+    new ReactRefreshWebpackPlugin({
+      disableRefreshCheck: true,
+    })
 
   /**
    * Extracts css requires into a single file;