+++ www/src/components/prev-and-next.js
@@ -1,5 +1,7 @@
 /** @jsx jsx */
 import { jsx } from "theme-ui"
+import { t, Trans } from "@lingui/macro"
+import { withI18n } from "@lingui/react"
 import Link from "./localized-link"
 import {
   MdArrowBack as ArrowBackIcon,
@@ -27,14 +29,14 @@ const prevNextLabelStyles = {
   mt: 0,
 }
 
-const PrevAndNext = ({ prev = null, next = null, ...props }) => {
+const PrevAndNext = ({ prev = null, next = null, i18n, ...props }) => {
   if (!prev && !next) {
     return null
   }
 
   return (
     <nav
-      aria-label="pagination"
+      aria-label={i18n._(t`pagination`)}
       sx={{
         [mediaQueries.sm]: {
           display: `flex`,
@@ -47,7 +49,9 @@ const PrevAndNext = ({ prev = null, next = null, ...props }) => {
       <div css={{ [mediaQueries.sm]: { width: `48%` } }}>
         {prev && (
           <Link to={prev.link} sx={prevNextLinkStyles}>
-            <p sx={prevNextLabelStyles}>Previous</p>
+            <p sx={prevNextLabelStyles}>
+              <Trans>Previous</Trans>
+            </p>
             <span
               sx={{
                 [mediaQueries.md]: {
@@ -78,7 +82,9 @@ const PrevAndNext = ({ prev = null, next = null, ...props }) => {
       >
         {next && (
           <Link to={next.link} sx={prevNextLinkStyles}>
-            <p sx={prevNextLabelStyles}>Next</p>
+            <p sx={prevNextLabelStyles}>
+              <Trans>Next</Trans>
+            </p>
             <span
               sx={{
                 [mediaQueries.md]: {
@@ -104,4 +110,4 @@ const PrevAndNext = ({ prev = null, next = null, ...props }) => {
   )
 }
 
-export default PrevAndNext
+export default withI18n()(PrevAndNext)